backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.13
alembic>=1.14.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
google-auth>=2.37.0
google-genai>=1.29.0
openai>=1.0.0
httpx>=0.28.1
pytest>=8.3.4
pytest-asyncio>=0.25.0
